+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     * Create a New Magento 2 Module action
     * Code Inspection: Duplicated Observer Usage in events XML
     * Create a Plugin class for a class public method action
+    * Code Inspection: Warning regarding Cacheable false attribute in default XML 
 
 0.3.0
 =============